Solr Configuration Revision

+++ src/main/java/n11/n11finalcaseahmetsaimyilmaz/restaurant/RestaurantRepository.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import java.util.Optional;
 
 @Repository
-public interface RestaurantRepository extends SolrCrudRepository<Restaurant, Long> {
+public interface RestaurantRepository extends SolrCrudRepository<Restaurant, Integer> {
 
 
 }
